made nan_fill_dot more memory efficient

+++ sklearn/metrics/pairwise.py
@@ -461,7 +461,7 @@ def _nan_fill_dot(X, Y, fill_values: np.ndarray):
     if X.shape[1] != fill_values.shape[0]:
         raise ValueError('X and fill_values have incompatible shapes')
 
-    p = np.zeros((X.shape[0], Y.shape[0],))
+    p = np.zeros((X.shape[0], Y.shape[0],), dtype=np.float32)
     for i in prange(X.shape[0]):
         v1 = X[i, :]
         for j in range(Y.shape[0]):
@@ -486,7 +486,7 @@ def nan_filled_euclidean_distances(X, fill_values, Y=None, squared=False, copy=T
         YY = YY[np.newaxis, :]
 
     # if dtype is already float64, no need to chunk and upcast
-    distances = - 2 * _nan_fill_dot(X, Y, -fill_values)
+    distances = - 2 * _nan_fill_dot(X.astype(np.float32), Y.astype(np.float32), -fill_values.astype(np.float32))
     # distances = - 2 * np.dot(X, Y.T)
     distances += XX
     distances += YY
